refactor(changes): isolate row presentation
